fix: delete usePersistedFlags hook. Move code to persistedFlagsSlice.
