+++ cardstack/src/screens/Profile/ProfileNameScreen/useProfileNameScreen.ts
@@ -4,7 +4,7 @@ import { useState, useCallback, useMemo } from 'react';
 import { useProfileUpdate } from '@cardstack/hooks';
 import { Routes } from '@cardstack/navigation';
 import { RouteType } from '@cardstack/navigation/types';
-import { usePersistedFlags } from '@cardstack/redux/hooks/usePersistedFlags';
+import { usePersistedFlagsActions } from '@cardstack/redux/persistedFlagsSlice';
 import { CreateProfileInfoParams } from '@cardstack/services/hub/hub-types';
 import { MerchantInformation } from '@cardstack/types';
 import { contrastingTextColor } from '@cardstack/utils';
@@ -73,7 +73,7 @@ export const useProfileNameScreen = () => {
     updateProfile({ ...profile, id: currentProfile.id || '' });
   }, [navigate, profile, currentProfile, updateProfile]);
 
-  const { triggerSkipProfileCreation } = usePersistedFlags();
+  const { triggerSkipProfileCreation } = usePersistedFlagsActions();
 
   return {
     onContinuePress,
