+++ cardstack/src/screens/Profile/ProfileSlugScreen/useProfileSlugScreen.ts
@@ -4,7 +4,7 @@ import { useState, useCallback, useEffect, useMemo, useRef } from 'react';
 
 import { useInitIAPProducts } from '@cardstack/hooks/usePurchaseProfile';
 import { Routes } from '@cardstack/navigation';
-import { usePersistedFlags } from '@cardstack/redux/hooks/usePersistedFlags';
+import { usePersistedFlagsActions } from '@cardstack/redux/persistedFlagsSlice';
 import { useLazyValidateProfileSlugQuery } from '@cardstack/services';
 import { matchMinLength } from '@cardstack/utils/validators';
 
@@ -80,7 +80,7 @@ export const useProfileSlugScreen = () => {
     [isFetching, isLoading, slugValidation.slugAvailable]
   );
 
-  const { triggerSkipProfileCreation } = usePersistedFlags();
+  const { triggerSkipProfileCreation } = usePersistedFlagsActions();
 
   return {
     slug,
