+++ cardstack/src/screens/Profile/PurchaseCTAScreen/usePurchaseCTAScreen.ts
@@ -6,7 +6,7 @@ import { useCreateProfile } from '@cardstack/hooks/merchant/useCreateProfile';
 import { usePurchaseProfile } from '@cardstack/hooks/usePurchaseProfile';
 import { Routes } from '@cardstack/navigation';
 import { RouteType } from '@cardstack/navigation/types';
-import { usePersistedFlags } from '@cardstack/redux/hooks/usePersistedFlags';
+import { usePersistedFlagsActions } from '@cardstack/redux/persistedFlagsSlice';
 import { useGetSafesDataQuery } from '@cardstack/services';
 import { CreateProfileInfoParams } from '@cardstack/services/hub/hub-types';
 import { remoteFlags } from '@cardstack/services/remote-config';
@@ -74,7 +74,7 @@ export const usePurchaseCTAScreen = () => {
     setPurchaseEnd();
   }, [purchaseProfile, setPurchaseStart, setPurchaseEnd]);
 
-  const { triggerSkipProfileCreation } = usePersistedFlags();
+  const { triggerSkipProfileCreation } = usePersistedFlagsActions();
 
   return {
     onPressChargeExplanation,
